Move logging to its own package, allowing us to singleton it
