+++ inject/injector.go
@@ -4,14 +4,15 @@ package inject
 
 import (
 	"context"
+	logger "logging"
 
 	"cloud.google.com/go/pubsub"
 	"github.com/ONSdigital/ras-rm-sample/file-uploader/config"
 	"github.com/ONSdigital/ras-rm-sample/file-uploader/file"
 	"github.com/google/wire"
 	"github.com/spf13/viper"
-
-	log "github.com/sirupsen/logrus"
+	"go.uber.org/zap"
+	"gorm.io/gorm/logger"
 )
 
 var FileProcessor = Inject()
@@ -23,7 +24,6 @@ func Inject() file.FileProcessor {
 
 func ConfigSetup() config.Config {
 	viper.AutomaticEnv()
-	configureLogging()
 	viper.SetDefault("PORT", "8080")
 	viper.SetDefault("GOOGLE_CLOUD_PROJECT", "rm-ras-sandbox")
 	viper.SetDefault("PUBSUB_TOPIC", "topic")
@@ -41,16 +41,6 @@ func ConfigSetup() config.Config {
 	return config
 }
 
-func configureLogging() {
-	verbose := viper.GetBool("VERBOSE")
-	log.SetFormatter(&log.JSONFormatter{})
-	if verbose {
-		log.SetLevel(log.DebugLevel)
-	} else {
-		log.SetLevel(log.InfoLevel)
-	}
-}
-
 func NewFileProcessor(config config.Config, client *pubsub.Client, ctx context.Context) file.FileProcessor {
 	return file.FileProcessor{Config: config, Client: client, Ctx: ctx}
 }
@@ -62,8 +52,8 @@ func GenContext() context.Context {
 func NewPubSub(config config.Config, ctx context.Context) *pubsub.Client {
 	client, err := pubsub.NewClient(ctx, config.Pubsub.ProjectId)
 	if err != nil {
-		log.WithError(err).Error("Failed to create pubsub client")
+		logger.Error("Failed to create pubsub client", zap.Error(err))
 	}
-	log.Info("Pubsub client created")
+	logger.Info("Pubsub client created")
 	return client
 }
