+++ main.go
@@ -1,20 +1,12 @@
 package main
 
 import (
-	"github.com/ONSdigital/ras-rm-sample/file-uploader/routes"
-	"github.com/blendle/zapdriver"
-	"go.uber.org/zap"
 	"net/http"
+
+	"github.com/ONSdigital/ras-rm-sample/file-uploader/routes"
 
 	"github.com/gorilla/mux"
 )
-
-var logger *zap.Logger
-
-func init() {
-	logger, _ = zapdriver.NewProduction()
-	defer logger.Sync()
-}
 
 func main() {
 	router := mux.NewRouter().StrictSlash(true)
